pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astromodal"
version = "0.1.0"
description = "Foundation model for astronomy - transformer-based multi-modal learning"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Astromodal Team"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "polars>=0.19",
    "torch>=2.0",
    "scikit-learn>=1.3",
    "astropy>=5.3",
    "healpy>=1.16",
    "pyarrow>=13.0",
    "gaiaxpy>=2.0",
    "tqdm>=4.65",
    "joblib>=1.3",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "black>=23.0",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
astromodal = ["config/*.yaml"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"
